dedicated track component (always rendered in app)
